requirements.txt
numpy>=1.21.2
opencv-python>=4.7.0
mediapipe>=0.10.0
flask>=2.0.1,<3.0.0
flask-socketio>=5.1.1,<6.0.0
flask-cors>=3.0.10,<4.0.0
python-dotenv>=0.19.0
gevent-websocket>=0.10.1
pymongo>=3.12.0,<4.0.0
werkzeug>=2.0.0,<3.0.0
python-engineio>=4.0.0
python-socketio>=5.0.2
gunicorn>=20.1.0 